Split scatteringtheory into 2 classes, an ImageFormation class which does image formation, and a ScatteringTheory class which just calculates scattering
